adding new update for stutruce and files and move getReqeuset to new file and slove the problem for showpofileuser in comment
