+++ index.html
@@ -232,6 +232,8 @@
     <!--// adding loading for the pagination // -->
   </div>
   <!-- // start post //  -->
+
+  <!-- the main script file -->
   <script type="module" src="script.mjs"></script>
 </body>
