+++ src/js/Logic/handleShowInfoUser.mjs
@@ -18,7 +18,7 @@ const showUserInfo = async (element, foundUserId) => {
   } else {
     // if(foundUserId == null) {
     user = JSON.parse(decodeURIComponent(element));
-    console.log(user)
+    // console.log(user)
     let user_id = user.author.id;
     userProfile = false;
     // console.log(user)
