+++ src/js/templates/commentTemplate.mjs
@@ -1,9 +1,7 @@
 
-function templateComment(item, postInfo, conditionComment, user_id) {
-  window.user_id = user_id; // to use it in the showUserInfo function to be global in the window object
-  console.log(conditionComment)
+function templateComment(item) {
   const comm = `
-    <div class="comment pb-1" onclick={showUserInfo('${encodeURIComponent( JSON.stringify(postInfo),)}')} >
+      <div class="comment pb-1" onclick={showUserInfo('${encodeURIComponent( JSON.stringify(item))}')} >
         <div class="pt-1 flex align-items-center justify-content-center gap-3 pb-1">
           ${typeof item.author.profile_image === "string" ?
       `<img src="${item.author.profile_image}" alt="Avatar1" class="img-fluid my-1" style="width: 30px; height: 30px; border-radius: 50% !important;" />`
